forall: continue on errors
Mostly motivated by `./everest forall git grep blah`. If one of the
greps comes out empty, it also has a non-zero exit code and the script
will not run the command on the remaining repos (since we have -e). The
same happens when git invokes a pager (less) and one presses 'q' to
quit.

While at it use a subshell instead of cd-ing back to the parent
directory.