everest: bootstrap F*, now required for building Pulse
Pulse relies on checked files for internal compiler modules, after
https://github.com/FStarLang/pulse/pull/246.

We use to avoid this need by keeping an OCaml snapshot of the checker in
the Pulse repo, but we've removed that since. That was not ideal due
to merge conflicts and etc, but also meant that an everest build would
not check that Pulse can successfully build.

This changes the F* build rule to add the bootstrapping step.